pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tenfabric"
version = "0.1.0"
description = "One command to provision, train, and export fine-tuned language models"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10,<3.13"
authors = [{ name = "Ramkumar", email = "ram@billaram.dev" }]
keywords = ["llm", "fine-tuning", "training", "infrastructure", "slm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "typer>=0.12",
    "pydantic>=2.0",
    "rich>=13.0",
    "pyyaml>=6.0",
    "temporalio>=1.7",
    "httpx>=0.27",
]

[project.optional-dependencies]
training = [
    "torch>=2.0",
    "transformers>=4.40",
    "trl>=0.29",
    "datasets>=2.18",
    "peft>=0.11",
    "accelerate>=0.30",
    "bitsandbytes>=0.43",
]
unsloth = [
    "unsloth>=2024.0",
]
# Base skypilot (no cloud provider credentials baked in)
cloud = [
    "skypilot>=0.9",
]
# Individual cloud providers — install the ones you use
cloud-aws = [
    "skypilot[aws]>=0.9",
]
cloud-gcp = [
    "skypilot[gcp]>=0.9",
]
# NOTE: azure extra is excluded from uv resolution due to broken azure-cli deps.
# Install manually if needed: uv pip install 'skypilot[azure]' --prerelease=allow

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
tfab = "tenfabric.cli.app:app"

[project.urls]
Homepage = "https://github.com/billaram/tenfabric"
Documentation = "https://tenfabric.dev"
Repository = "https://github.com/billaram/tenfabric"
Issues = "https://github.com/billaram/tenfabric/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/tenfabric"]

[tool.uv]
# Only resolve for Linux and macOS — skypilot doesn't support Windows
environments = [
    "sys_platform == 'linux'",
    "sys_platform == 'darwin'",
]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.10"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
